gpu: nvgpu: clk: update type for size param
Update interfaces to use size_t to align with boardobj_construct_super()
and avoid unnecessary casts.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2011434
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/pmu/clk/clk_fll.c
@@ -430,7 +430,7 @@ static struct fll_device *construct_fll_device(struct gk20a *g,
 
 	nvgpu_log_info(g, " ");
 	status = boardobj_construct_super(g, &board_obj_ptr,
-		(u32)sizeof(struct fll_device), pargs);
+		sizeof(struct fll_device), pargs);
 	if (status != 0) {
 		return NULL;
 	}
